+++ src/combo_selector/ui/pages/export_page.py
@@ -37,7 +37,7 @@ from combo_selector.ui.widgets.flat_radio_grouped_button import FlatRadioGrouped
 from combo_selector.ui.widgets.line_widget import LineWidget
 from combo_selector.ui.widgets.neumorphism import BoxShadow
 from combo_selector.ui.widgets.style_table import StyledTable
-from combo_selector.utils import resource_path
+from combo_selector.resource_utils import resource_path
 
 # Get icon path for dropdown arrow
 drop_down_icon_path = resource_path("icons/drop_down_arrow.png").replace("\\", "/")