+++ src/combo_selector/ui/pages/om_calculation_page.py
@@ -40,7 +40,7 @@ from combo_selector.ui.widgets.neumorphism import BoxShadow
 from combo_selector.ui.widgets.style_table import StyledTable
 from combo_selector.ui.widgets.section_help_button import SectionHelpButton
 from combo_selector.ui.widgets.animated_toogle_button import AnimatedSegmentedToggle
-from combo_selector.utils import resource_path
+from combo_selector.resource_utils import resource_path
 from combo_selector.constants import METRIC_PLOT_MAP
 
 # Dropdown arrow icon path