+++ src/combo_selector/ui/pages/redundancy_check_page.py
@@ -39,7 +39,7 @@ from combo_selector.ui.widgets.neumorphism import BoxShadow
 from combo_selector.ui.widgets.qcombobox_cmap import QComboBoxCmap
 from combo_selector.ui.widgets.style_table import StyledTable
 from combo_selector.ui.widgets.section_help_button import SectionHelpButton
-from combo_selector.utils import resource_path
+from combo_selector.resource_utils import resource_path
 from combo_selector.constants import METRIC_CORR_MAP
 from combo_selector.core.orthogonality_utils import build_correlation_matrix_for_display